+++ bin/nanocld/changelog.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [untagged]
 
-### Feature
+### Added
 
+- Endpoint to restart a cargo
+- StateFile use Kind instead of Type
 - Replication `Static` that can allow development tests
 
 ### Fixed